+++ Xendit/M2Invoice/Controller/Checkout/ProcessHosted.php
@@ -26,6 +26,16 @@ class ProcessHosted extends AbstractAction
                     return $this->handlePaymentFailure($order, $hostedPayment['error_code'], 'Error reconciliating');
                 }
 
+                if ($hostedPayment['paid_amount'] != $hostedPayment['amount']) {
+                    $order->setBaseDiscountAmount($hostedPayment['paid_amount'] - $hostedPayment['amount']);
+                    $order->setDiscountAmount($hostedPayment['paid_amount'] - $hostedPayment['amount']);
+                    $order->save();
+    
+                    $order->setBaseGrandTotal($order->getBaseGrandTotal() + $order->getBaseDiscountAmount());
+                    $order->setGrandTotal($order->getGrandTotal() + $order->getDiscountAmount());
+                    $order->save();
+                }
+
                 return $this->processSuccessfulTransaction(
                     $order,
                     $payment,